Updates
RegionMask Indicators have been updated so that they dont fail when trying to buffer an area with no indications

Also convertMask has been given the 'shrink' input which shrink all the produced geometries by a tiny amount. This is done in order to prevent the creation of invalid geometries (i.e. havine self-intersections)